+++ R/utils.R
@@ -132,6 +132,19 @@ hasPrefix <- function(x, prefix) {
   substring(x, 1, nchar(prefix)) == prefix
 }
 
+# Lightweight equivalent of withr::defer()
+defer <- function(expr, env = caller_env(), after = FALSE) {
+  thunk <- as.call(list(function() expr))
+  do.call(on.exit, list(thunk, TRUE, after), envir = env)
+}
+
+dirCreate <- function(paths) {
+  for (path in paths) {
+    dir.create(path, showWarnings = FALSE, recursive = TRUE)
+  }
+  paths
+}
+
 fromIDE <- function() {
   !is.na(Sys.getenv("RSTUDIO", unset = NA)) && !identical(.Platform$GUI, "RStudio")
 }