+++ tests/testthat/test-writeManifest.R
@@ -1,5 +1,5 @@
 makeManifest <- function(appDir, appPrimaryDoc = NULL, ...) {
-  writeManifest(appDir, appPrimaryDoc = appPrimaryDoc, ...)
+  writeManifest(appDir, appPrimaryDoc = appPrimaryDoc, ..., quiet = TRUE)
   manifestFile <- file.path(appDir, "manifest.json")
   data <- readLines(manifestFile, warn = FALSE, encoding = "UTF-8")
   manifestJson <- jsonlite::fromJSON(data)