+++ NEWS.md
@@ -10,6 +10,22 @@
 * A `version` field has been added to deployment DCF files to facilitate file
   format changes. Its value for this release is `1`.
 
+* `deploySite()` now supports quarto websites (#813).
+
+* `deployApp()` and `writeManifest()` now respect renv lock files, if present. 
+  If you don't want to use these lockfiles, and instead return the previous 
+  behaviour of snapshotting on every deploy, add your `renv.lock` to 
+  `.rscignore` (#671). Learn more `?appDependencies()`.
+  
+* `deployApp()` and `writeManifest()` now use renv to capture app dependencies, 
+  rather than packrat. If this causes a previously working deploy to fail, 
+  please file an issue then set `options(rsconnect.packrat = TRUE)` to revert 
+  to the previous behaviour.
+
+* The built-in linter should have fewer false positives for path problems:
+  the relative path linter has been removed (#244) and the case-sensitive 
+  linter now only checks strings containing a `/` (#611).
+
 * `deployApp()`'s `quarto` argument now takes values `TRUE`, `FALSE` or 
   `NA`. The previous value (a path to a quarto binary) is now deprecated,
   and instead we automatically figure out the packge from `QUARTO_PATH` and